@@ -7,10 +7,10 @@ import 'dart:math';
 class GradientCircularProgressIndicator extends StatelessWidget {
   final double strokeWidth;
   final bool strokeRound;
-  final double value;
+  final double? value;
   final Color backgroundColor;
   final List<Color> gradientColors;
-  final List<double> gradientStops;
+  final List<double>? gradientStops;
   final double radius;
 
   /// Constructor require progress [radius] & gradient color range [gradientColors]
@@ -21,8 +21,8 @@ class GradientCircularProgressIndicator extends StatelessWidget {
   GradientCircularProgressIndicator({
     this.strokeWidth = 10.0,
     this.strokeRound = false,
-    @required this.radius,
-    @required this.gradientColors,
+    required this.radius,
+    required this.gradientColors,
     this.gradientStops,
     this.backgroundColor = Colors.transparent,
     this.value,
@@ -36,10 +36,8 @@ class GradientCircularProgressIndicator extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     var _colors = gradientColors;
-    if (_colors == null) {
-      Color color = Theme.of(context).accentColor;
-      _colors = [color, color];
-    }
+    Color color = Theme.of(context).accentColor;
+    _colors = [color, color];
 
     return Transform.rotate(
       angle: -pi / 2,
@@ -69,31 +67,31 @@ class _GradientCircularProgressPainter extends CustomPainter {
     this.radius,
   });
 
-  final double strokeWidth;
-  final bool strokeRound;
-  final double value;
+  final double? strokeWidth;
+  final bool? strokeRound;
+  final double? value;
   final Color backgroundColor;
-  final List<Color> gradientColors;
-  final List<double> gradientStops;
+  final List<Color>? gradientColors;
+  final List<double>? gradientStops;
   final double total;
-  final double radius;
+  final double? radius;
 
   @override
   void paint(Canvas canvas, Size size) {
     if (radius != null) {
-      size = Size.fromRadius(radius);
+      size = Size.fromRadius(radius!);
     }
 
     double _value = (value ?? .0);
     _value = _value.clamp(.0, 1.0) * total;
     double _start = .0;
 
-    double _offset = strokeWidth / 2;
+    double _offset = strokeWidth! / 2;
     Rect rect = Offset(_offset, _offset) &
-        Size(size.width - strokeWidth, size.height - strokeWidth);
+        Size(size.width - strokeWidth!, size.height - strokeWidth!);
 
     var paint = Paint()
-      ..strokeWidth = strokeWidth
+      ..strokeWidth = strokeWidth!
       ..style = PaintingStyle.stroke
       ..isAntiAlias = true;
 
@@ -104,7 +102,7 @@ class _GradientCircularProgressPainter extends CustomPainter {
 
     if (_value > 0) {
       paint.shader = SweepGradient(
-              colors: gradientColors,
+              colors: gradientColors!,
               startAngle: 0.0,
               endAngle: _value,
               stops: gradientStops)
